fixed some extra warnings

+++ include/rand.h
@@ -8,7 +8,7 @@
 
 #define rng_init()  srand(time(0))
 #define rnd(a,b)    dice_roll((a), (b))
-#define rn(n)       (rand()%(n))
+#define rn(n)       ((int)(rand()%(n)))
 
 /* Random number, simulating rolling a dice with b faces (sides). NOT EVEN -
    biased towards the middle range. */
